Refractored pass gen and fixed while loop cases for prompt cancel selection. Fixed icons in weather and adjusted 5 day to only display 12pm for each day.

+++ src/components/weather/components/Search.jsx
@@ -26,9 +26,14 @@ function Search({getWeather, searchList}) {
   },[]);
 
   const updateSearchHistory = (city) => {
-    const updatedSearchList = [city, ...searchList];
-    localStorage.setItem("searchHistory", JSON.stringify(updatedSearchList));
+    const existingCities = JSON.parse(localStorage.getItem("searchHistory")) || [];
+
+    if(!existingCities.includes(city)){
+      const updatedSearchList = [city, ...existingCities];
+      localStorage.setItem("searchHistory", JSON.stringify(updatedSearchList));
+    }
   };
+
   return (
     <div className="container">
       <h3>Search for a City:</h3>
